Have raid-discover bind raid-configure into the noderole graph.
Currently, this just binds it straight into the system deployment and
does not give the user a change to customize the RAID config before it
is applied to a node.  There is design work needed to enable that --
this is mainly here to verify that the raid configure recipes still work.