Fix BMC presence detection on F20.
Unlike every other userspace, Fedora 20 does not present DMI data by
default.  Since that is the case, flesh out the function that sees if
IPMI is supported on the node to fail of the node does not have any DMI
information.

This may or may not be the right way to do it.