Add bios-discover and bios-dell-rseries-configure roles.
The bios-discover role uses node attributes to match against BIOS config
group selectors stored in its template data.  Based on what the best
match is, it binds the appropriate bios-configure role in
place (currently, the only one implemented is
bios-dell-rseries-configure), which is responsible for actually
configuring the BIOS on the system.