Change where and how we set the admin hostname.
The crowbar-* scripts are no longer responsible for setting the
admin node's hostname.  Instead, that is the job of production.sh.

Goiardi has also been updated to explicitly use the hostname of the node
at the time it is initialized, anstead of using whatever hostname the
node happens to have when the process is started.