Micro-service to manage DNS
Adds a go-based program that runs a micro-service to manage
DNS servers (BIND or PowerDNS).  The Crowbar-API/Runner Services
call the micro-service to update DNS mappings as nodes receive or
lose addresses or change names.

The system assigns addresses per network assigned.  The network has
parameters for domain and hostname templates.  These are used to
generate a set of updates to the DNS server.

The micro-service presents a SSL-based access point that allows
for updates to the DNS server.  For BIND, it will rewrite config
files.  For PowerDNS, the PowerDNS API will be used to update.
Zones will be created if they don't exist.