Handle DHCP handoff a bit better.
There was apparently a brief window while between killing DHCP and
adding the static address that let us believe that we had added the
static address but we were still running on the dynamic address.

Logging has been added to see when this happens, along with a 3 second
time out and better flushing when we clear out the addresses of the nic
we are taking over.