Add session timer for ajax calls so that we don't run sessions forever even if the page is polling.
This approach relies on the server to track session information
because the page refreshs make it hard to persist long term
values in the browser.

Basically, the application layout hearbeat has a hidden
span value that holds the session length in seconds.

All of the screen refresh pages have been updated to check
that value before they do a refresh.  It has been hardcoded
at an hour.

You can inspect the value of the timer by turning on
the edge UI value.  that will show the session seconds in
the top right corner with a clock icon.

Connnect to #756

Signed-off-by: robhirschfeld <rob@zehicle.com>