Use the Go CLI instead of the Ruby one for bootstrapping.
The Go CLI can now handle everything needed for bootstrapping except for
barclamp importing, which (due to marshalling incompatibilities between
YAML and JSON on the Go side) is still handled by a stub Ruby program.

Since the Go CLI is a statically linked binary, it can be used as-is by
basically every 64 bit Linux system without having to deal with gems and
the whole Ruby infrastructure.