Make after_commit on: :create hooks run once and only once.
If you modify an object in an after_commit on: :create hook, it reruns
the hook despite that not being the logical thing to do (because the
object has already been created saved to the database, and the
transaction in which it was created has already ended).

Thanks for the weird behaviour again, Rails.