Dynamically build docker-compose.yml file to meet the needs of the deployment.
./init_file.sh takes args to edit the default settings.

--clean - removes dynamic files and exits 0

--provisioner - adds the provisioner components and adds an admin network
--logging - adds kibana, elasticsearch, logspout, and logstash
--debug - adds cadvisor for container debugging
--node - adds node container that acts as a joined OS.

--access [HOST|FORWARDER] - changes the mode that the admin node works.

HOST mode means that an IP of the HOST system will be used as the access point
into the admin node.  This is set as the EXTERNAL_IP in the init_files.sh.
[Add flag to override this]

FORWARDER mode means that a forwarder container will handle access.  This is
specified by the FORWARDER_IP (and the EXTERNAL_IP should be set to match it).

FORWARDER mode is most like our dev modes where we bridge an interface into
docker0 and run the admin network on that interface.  The default admin network
and FORWARDER_IP value cause the forwarder to be in the admin network and assumes
some bridging is happening either a physical port is bridged into the
docker0 bridge or kvm-slaves/docker containers are used as the nodes already
in docker0 bridge space.

The vagrant and ansible playbooks have been updated to use the defaults in
init_file.sh which are FORWARDER mode and no provisioner.
[May want to change this to have the provisioner]