Have Consul always run in host mode.
This modifies the templates and init_files.sh to allow the Consul
container to always run in host mode.  It tries to be smart about
telling the Consul server what address on the host to advertise to the
other consuls, and complains and dies loudly if it cannot pick one.