Add a smart container rebuilding program.
rebuild-containers is a smart container rebuilding script.  It
understands how the containers relate to each other, and can figure
out (via container build tracking branches and container image IDs
what containers would need to be rebuilt when any of them change.

It takes a couple of options:

   --tag if you want to tag the built containers with something other
     than latest

   --pull if you want to pull containers from the Docker hub before
     being built

Anything else is interpreted as either the name of a container to be
built, or a path to a container repository.