Make rebar_api startup a bit smarter.
This patch updates init_files.sh to pass a list of services that
rebar_api should wait to be bound in the system deployment before
letting the annealer run.  It also rearranges the startup scrupt
ordering to ensure that we only run initial bootstrap tasks after all
the services we require are present and accouted for.