Allow parallel client auth paths for rebar cli.
The rebar client will auth to the admin endpoint using
the username and password as normal.

The response to the initial ping will detect if the auth method
is digest or saml.  If digest, the digest will be parsed and
stored for use on subsequent requests.

If the saml, the returned headers contain form data (pre-processed
by the rev-proxy to avoid more form parsing) that can be used to build
an auth request (OKTA specific?) to the IDP to get a onetime sign in token.
This token is used to generate forward the SAML auth request to be
signed and validated by the IDP.  This returns a form that is parsed
and posted back to the DR admin.  Once validated, the DR admin returns
a DR token to use in subsequent requests.  Whew!

Also includes fixes to make sure that the resp bodies are cleaned up.