Fix up where a couple of noderoles are added.
amt-configure should always be added in the system deployment so that it
can be used to power off nodes that have nothing else to do.

raid-discover now adds the noderoles it adds on sucessful discovery to
the deployment the node is currently a member of, instead of the
deployment the discovery noderole is in.