Improve locking around runs to increase parallelism.
Locking around the runs table was being a little too paranoid, resulting
in the runners being starved to work due to the overhead in acquiring
the right locks.  This commit adds a partial index to guarantee that we
can only ever have one run per node in the database, relaxes some
locking that was previously enforcing that invariant, and takes
advantage of new FOR UPDATE SKIP LOCKED in Postgres 9.5 that is designed
to help highly contended queue-like tables perform better.