Add trusted session support into the rebar API and CLI.
Trusted sessions are to be used when operating in the trust zone -- in
this mode, all communications with the API server are authenticated by
TLS mutual authentication, and the client can effectively act on behalf
of any user.  A trusted session can only be created where the client can
get access to the trust-me service and have it issue an ephemeral
certificate.