Update DHCP code to handle lease code a little more gracefully.
If the DHCP address range for a subnet changes, don't refuse to issue
any more addresses until the current leases expire.  Instead, mark any
leases that are out of the current address range as invalid and NAK them
the next time the client tries to update its lease.

While we are at it, implement phantom leases to track when clients NAK a
lease on an address.  These phantom leases will blacklist the IP we
tried to hand out for 30 seconds.