Add build-release.sh and tweak up infrastructure.
Several changes here:

Rename Dockerfiles to Dockerfile.in to allow for tweaking FROM lines.
When building releases, we want to ensure that containers are built
using the proper tags throughout, but docker build does not allow for
that to happen natively.  Instead, generate and remove Dockerfiles as
needed.

Fix up all the places we use DR_TAG to set it to the name of the current
branch by default unless explicitly set in the environment or elsewhere.

Tweak rebuild-containers and build-rebar.sh to operate with build-release.sh.