During node_role create cohort wasn't getting set unless the role was a member of a cluster.  One problem is that parent test vs current cohort value was < and not <=.  We always want the new node role's cohort to be at least one more that all of its parents.
This also updates the cohort on preceed additions to the graph.
If the new role is a parent to roles (because of preceeds), then
the cohort of the children need to be updated.
If the new role is a child a preceeds parent, then we make sure the
new cohort value for the new child is larger than the parent.

This seems to take care of most of the delete, remove, reset issues
I was seeing.  I'm not sure when this broke.