Add profile support to Rebar
This adds profiles as new top-level constructs in Digital Rebar.

A profile is a named mapping of attrib names to default values for those
attribs.  Unless overridden by the user (by setting attribs via the UI,
API, or CLI), attribs in a profile will take precedence over default
attribs from roles and deployment roles.

Nodes have been updated to contain a new top-level field called
'profiles', which should contain a list of profiles that apply to the
node.  Profiles apply to the node from the last to the first in the list
-- attrib values defined in a profile towards the fromt of the list get
merged in with attrib values from later in the list.  The logic is
similar to how attribs get smashed together -- the default provide the
basics, which the deploymentroles/noderoles/nodes can override, with the
most specific ones taking precendce over the less specific ones.