First step in some reconciliation.
This adds a history variable that tracks updates over time.
This adds the wanted_version variable that can be latest or a specific
version.  If the value is latest, the code will walk from youngest to
oldest applying the updates in order until the version is the "latest"
in the list. If a specific version is specified, only that version will
be specified.  If the version is less than the current version, a force
variable is set and the scripts are updated to use it.

One design choice that was punted for now is if the wanted_version
is say 10, and the current is 6 and we know about 8 and 10, we will
only apply 10.  There are cases where this is desired and then there
are cases where we want to apply 8 and then 10.