Update the attribute update function to poke the childern of node roles of deployment roles.  Also make a copy of the original structure so that when the update function works we can actually see if the value changes.
The poke function needs to force the node roles to block or todo.
The blocking state will wait until committed data is commited on the
deployment or node role when they are committed.  Otherwise, we
either don't change the state of the node roles or we run the
node roles with old data.

The FLOW for updating attributes is, mark the deployment proposed,
update the attribute, commit the deployment.  This will generate the
proper running of node roles with the correct data.