Add a generic interface-based save and load set for SimpleStore
This adds a set of interfaces that objects can satisfy (along with
being marshallable to JSON) to enable them to be easily saved and
restored to a SimpleStore.  There are also additional interfaces that
they can satisfy to allow for hooks to be called at strategic points
in the object's lifecycle.