Allow leases and bindings in the same nework range.
This change allows Rebar to hand out both short-term leases and bind
reservations from the same network range.  The use case is where Rebar
is managing DHCP for a subnet that is expected to be completly
populated (hence we cannot set aside a pool of addresses for initial
system discovery), and so we will need to arrange to allocate the
addresses a system used for initial discovery to that system on a
long-term basis.

Sledgehammer has also been updated to pass all of the MAC addresses
for discovered Ethernet devices at node create time, instead of just
the MAC that the system happened to boot from.  This gets rid of
spurious node creation when (due to forces beyond our control) the
system PXE boots from a different nic during the provision process.